requirements.txt
openai-whisper==20240930
edge-tts==7.0.1
pydub==0.25.1
sounddevice==0.5.1
scipy==1.15.2
python-dotenv==1.1.0
ttkbootstrap==1.12.0
PyQt5==5.15.11
SpeechRecognition==3.14.2
torch==2.6.0
numpy==2.2.5
